requirements.txt
nltk==3.6.6
numpy==1.22.0
rouge_score==0.1.2
scikit_learn==1.1.1
sentencepiece==0.1.96
strsimpy==0.2.1
summ_eval==0.892
tensorboard==2.9.0
tensorboardX==2.5
torch==1.13.1
tqdm==4.62.3
transformers==4.20.1
openai==0.27.0
python-dotenv==1.0.0
tiktoken==0.3.0
datasets==2.10.1
evaluate==0.4.0
protobuf==3.20.0
seaborn==0.12.2